src/main/test/java/com/aoc/y2024/ProblemDay9Test.java
package com.aoc.y2024;

import com.aoc.ProblemDayTest;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import static org.fest.assertions.Assertions.assertThat;

public class ProblemDay9Test implements ProblemDayTest {

    private ProblemDay9 problem;

    @BeforeEach
    public void before() {
        problem = new ProblemDay9();
    }

    @Test
    @Override
    public void testSolveA() {
        assertThat(problem.solveA()).isEqualTo(6301895872542L);
    }

    @Test
    @Override
    public void testSolveB() {
        assertThat(problem.solveB()).isEqualTo(6323761685944L);
    }

}
